// osa9/course-info/src/types/coursePart.ts
export interface CoursePartBase {
  name: string;
  exerciseCount: number;
}

export interface CoursePartWithDescription extends CoursePartBase{
  name: string;
  description: string;
}

export interface CoursePartOne extends CoursePartWithDescription {
  name: "Fundamentals";
}

export interface CoursePartTwo extends CoursePartBase {
  name: "Using props to pass data";
  groupProjectCount: number;
}

export interface CoursePartThree extends CoursePartWithDescription {
  name: "Deeper type usage";
  exerciseSubmissionLink: string;
}

export interface CoursePartFour extends CoursePartWithDescription {
  name: "Final part";
  grading: string;
}

export type CoursePart = CoursePartOne | CoursePartTwo |
  CoursePartThree | CoursePartFour;